refactor(grib1, grib2): both readers read through ByteSource like NetCDF does
Grib1Reader and Grib2Reader indexed a Vec<u8> that had to be the whole
file. They are now generic over a ByteSource, defaulting to Vec<u8>, so
from_bytes callers are unchanged.

Each decode entry point prefetches the sections it reads in one batch
and reads exactly those. The scan finds each edition-N indicator with
find_forward and walks a message's headers through one cursor bounded
to the message, so garbage costs windows rather than a read per byte
and no data section is read to find the next message. Section parsers
are handed what slicing the whole-file buffer used to give them, so
every error counts the same bytes. from_message_at reads one message at
a known offset without scanning, the path a sidecar index range takes.

byte_offset is u64 and the section ranges are ByteRange, since they
address a file rather than a buffer; pds_p1_offset returns u64. Both
crates re-export ByteRange and ByteSource. Grib1Reader::bytes stays for
the in-memory reader napi patches P1 through.

A before/after oracle over every GRIB fixture, plus padded, truncated
and byte-flipped copies of each, is identical across every entry point,
errors included (12,643 + 17,348 lines). tests/byte_source.rs in each
crate pins the prefetch/read plan, the scan's read bounds, a decode from
a source holding only one message's range, and short sources failing
cleanly. The wasm bundle grows 0.7% raw and is re-recorded.

The windowed reader this needs is the HDF5 one, hoisted from netcdf's
hdf5::source into fieldglass_core::bytes (FileCursor, read_at, read_up_to,
scan_windows), with read_exact, find_forward, FileCursor::within and
peek_up_to beside it. netcdf implements its Fields trait for core's cursor;
core gains no dependency.

+++ crates/fieldglass-grib1/tests/scanner.rs
@@ -16,7 +16,7 @@ fn scans_two_concatenated_messages() {
     let reader = Grib1Reader::from_bytes(buf).expect("two-message buffer parses");
     assert_eq!(reader.message_count(), 2);
     assert_eq!(reader.messages[0].byte_offset, 0);
-    assert_eq!(reader.messages[1].byte_offset, FIXTURE.len());
+    assert_eq!(reader.messages[1].byte_offset, FIXTURE.len() as u64);
 
     // Both messages decode to the same field.
     let a = reader.decode_message_values(0).expect("decode first");
@@ -32,7 +32,7 @@ fn skips_leading_non_grib_bytes() {
 
     let reader = Grib1Reader::from_bytes(buf).expect("parses past leading junk");
     assert_eq!(reader.message_count(), 1);
-    assert_eq!(reader.messages[0].byte_offset, prefix_len);
+    assert_eq!(reader.messages[0].byte_offset, prefix_len as u64);
 }
 
 #[test]
@@ -58,7 +58,7 @@ fn pds_p1_offset_round_trips_through_byte_buffer() {
     // — and patching that byte should be visible after re-parsing.
     let reader = Grib1Reader::from_bytes(FIXTURE.to_vec()).expect("fixture parses");
     let msg = &reader.messages[0];
-    let off = msg.pds_p1_offset();
+    let off = msg.pds_p1_offset() as usize;
     let original_p1 = msg.pds.p1;
     assert_eq!(
         FIXTURE[off], original_p1,
